+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Gov.hs
@@ -61,16 +61,16 @@ import Cardano.Ledger.Conway.Governance (
  )
 import Cardano.Ledger.Conway.Governance.Procedures (
   GovAction (..),
-  committeeMembersL,
  )
 import Cardano.Ledger.Conway.Governance.Snapshots (snapshotGovActionStates)
 import Cardano.Ledger.Conway.PParams (
   ConwayEraPParams (..),
   ppGovActionDepositL,
   ppGovActionExpirationL,
-  ppMinCommitteeSizeL,
  )
 import Cardano.Ledger.Core
+import Cardano.Ledger.Credential (Credential)
+import Cardano.Ledger.Keys (KeyRole (..))
 import Cardano.Ledger.Rules.ValidationMode (Inject (..), Test, runTest)
 import Cardano.Ledger.Shelley.Tx (TxId (..))
 import Control.DeepSeq (NFData)
@@ -91,7 +91,6 @@ import qualified Data.Set as Set
 import GHC.Generics (Generic)
 import Lens.Micro ((%~), (&), (^.))
 import NoThunks.Class (NoThunks (..))
-import Numeric.Natural (Natural)
 import Validation (failureUnless)
 
 data GovEnv era = GovEnv
@@ -105,11 +104,6 @@ data ConwayGovPredFailure era
   | MalformedProposal (GovAction era)
   | ProposalProcedureNetworkIdMismatch (RewardAcnt (EraCrypto era)) Network
   | TreasuryWithdrawalsNetworkIdMismatch (Set.Set (RewardAcnt (EraCrypto era))) Network
-  | NewCommitteeSizeTooSmall
-      -- | Size of `Committee` in `ProposalProcedure`
-      Natural
-      -- | `ppMinCommitteeSize` from `PParams`
-      Natural
   | ProposalDepositIncorrect
       -- | Submitted deposit
       Coin
@@ -119,6 +113,12 @@ data ConwayGovPredFailure era
     -- Voters. This failure lists all governance action ids with their respective voters
     -- that are not allowed to vote on those governance actions.
     DisallowedVoters !(Map.Map (GovActionId (EraCrypto era)) (Voter (EraCrypto era)))
+  | ConflictingCommitteeUpdate
+      -- | Credentials that are mentioned as members to be both removed and added
+      (Set.Set (Credential 'ColdCommitteeRole (EraCrypto era)))
+  | ExpirationEpochTooSmall
+      -- | Members for which the expiration epoch has already been reached
+      (Map.Map (Credential 'ColdCommitteeRole (EraCrypto era)) EpochNo)
   deriving (Eq, Show, Generic)
 
 instance EraPParams era => NFData (ConwayGovPredFailure era)
@@ -131,9 +131,10 @@ instance EraPParams era => DecCBOR (ConwayGovPredFailure era) where
     1 -> SumD MalformedProposal <! From
     2 -> SumD ProposalProcedureNetworkIdMismatch <! From <! From
     3 -> SumD TreasuryWithdrawalsNetworkIdMismatch <! From <! From
-    4 -> SumD NewCommitteeSizeTooSmall <! From <! From
-    5 -> SumD ProposalDepositIncorrect <! From <! From
-    6 -> SumD DisallowedVoters <! From
+    4 -> SumD ProposalDepositIncorrect <! From <! From
+    5 -> SumD DisallowedVoters <! From
+    6 -> SumD ConflictingCommitteeUpdate <! From
+    7 -> SumD ExpirationEpochTooSmall <! From
     k -> Invalid k
 
 instance EraPParams era => EncCBOR (ConwayGovPredFailure era) where
@@ -145,12 +146,14 @@ instance EraPParams era => EncCBOR (ConwayGovPredFailure era) where
         Sum ProposalProcedureNetworkIdMismatch 2 !> To acnt !> To nid
       TreasuryWithdrawalsNetworkIdMismatch acnts nid ->
         Sum TreasuryWithdrawalsNetworkIdMismatch 3 !> To acnts !> To nid
-      NewCommitteeSizeTooSmall submitted expected ->
-        Sum NewCommitteeSizeTooSmall 4 !> To submitted !> To expected
       ProposalDepositIncorrect submitted expected ->
-        Sum ProposalDepositIncorrect 5 !> To submitted !> To expected
+        Sum ProposalDepositIncorrect 4 !> To submitted !> To expected
       DisallowedVoters votes ->
-        Sum DisallowedVoters 6 !> To votes
+        Sum DisallowedVoters 5 !> To votes
+      ConflictingCommitteeUpdate members ->
+        Sum ConflictingCommitteeUpdate 6 !> To members
+      ExpirationEpochTooSmall members ->
+        Sum ExpirationEpochTooSmall 7 !> To members
 
 instance EraPParams era => ToCBOR (ConwayGovPredFailure era) where
   toCBOR = toEraCBOR @era
@@ -245,7 +248,7 @@ govTransition ::
   ConwayEraPParams era =>
   TransitionRule (ConwayGOV era)
 govTransition = do
-  TRC (GovEnv txid epoch pp, st, gp) <- judgmentContext
+  TRC (GovEnv txid currentEpoch pp, st, gp) <- judgmentContext
 
   let applyProps st' Empty = pure st'
       applyProps st' ((idx, ProposalProcedure {..}) :<| ps) = do
@@ -267,17 +270,24 @@ govTransition = do
                   Set.filter ((/= expectedNetworkId) . getRwdNetwork) $ Map.keysSet wdrls
              in Set.null mismatchedAccounts
                   ?! TreasuryWithdrawalsNetworkIdMismatch mismatchedAccounts expectedNetworkId
-          NewCommittee _mPrevGovActionId _oldColdCreds newCommittee ->
-            let minCommitteeSize = pp ^. ppMinCommitteeSizeL
-                newCommitteeSize = fromIntegral . Map.size $ newCommittee ^. committeeMembersL
-             in newCommitteeSize
-                  >= minCommitteeSize
-                    ?! NewCommitteeSizeTooSmall newCommitteeSize minCommitteeSize
+          UpdateCommittee _mPrevGovActionId membersToRemove membersToAdd _qrm -> do
+            checkConflictingUpdate
+            checkExpirationEpoch
+            where
+              checkConflictingUpdate =
+                let conflicting =
+                      Set.intersection
+                        (Map.keysSet membersToAdd)
+                        membersToRemove
+                 in Set.null conflicting ?! ConflictingCommitteeUpdate conflicting
+              checkExpirationEpoch =
+                let invalidMembers = Map.filter (<= currentEpoch) membersToAdd
+                 in Map.null invalidMembers ?! ExpirationEpochTooSmall invalidMembers
           _ -> pure ()
 
         let st'' =
               addAction
-                epoch
+                currentEpoch
                 (pp ^. ppGovActionExpirationL)
                 (GovActionId txid idx)
                 pProcDeposit