+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.9.0.0
 
+* Add `ExpirationEpochTooSmall` data constructor to `ConwayGovPredFailure`
+* Add `ConflictingCommitteeUpdate` data constructor to `ConwayGovPredFailure`
+* Rename `NewCommitte` to `UpdateCommittee`
+* Remove `NewCommitteeSizeTooSmall` data constructor from `ConwayGovPredFailure`
 * Fix invalid order in `fromGovActionStateSeq`, thus also `DecCBOR` for `ProposalsSnapshot`
 * Remove `DecCBOR`/`EncCBOR` and `FromCBOR`/`ToCBOR` for `RatifyState`, since that state
   is ephemeral and is never serialized.
@@ -51,6 +55,7 @@
 
 ### testlib
 
+* Rename `genNewCommittee` to `genUpdateCommitteee`
 * Add `genNewCommittee`
 * Add `genNoConfidence`
 * Add `genTreasuryWithdrawals`